# api/openapi.yaml
openapi: 3.1.0
info:
  title: go-unit-mangement API
  description: |
    JSON API of go-unit-mangement, served under `/api` next to the embedded web UI.

    Authentication uses a JWT access token sent as `Authorization: Bearer <token>`.
    It is returned by `POST /api/auth/login` or handed to the web UI at the end
    of the SSO flow, and is valid until it expires or its session is ended
    (sign-out, password reset, account deletion).

    When `OIDC_ACCESS_TOKEN_AUDIENCE` is set, a JWT access token issued by the
    SSO provider (or by an issuer in `OIDC_ACCESS_TOKEN_ISSUERS`) is accepted
    as the bearer token too, without signing in first. It must be signed with
    the issuer's published keys, unexpired, and carry one of the configured
    audiences in `aud`; ID tokens and opaque tokens are rejected. Its `sub`
    resolves to the linked account, which is created on first use.

    An RS256 JWT whose `kid` header names a registered API key
    (`POST /api/users/{id}/api-keys`) is accepted too and acts as that key's
    user. It must be signed with the key's private half, carry `iat` and
    `exp` no more than an hour apart, and be unexpired. Other instances
    syncing units from this one authenticate this way (see the `sync` tag).

    Errors are returned as `{"error": "<message>"}`.
    Request bodies are limited to 1 MiB.
  version: 1.0.0
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Sign in with username and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in; the response carries the access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid username or password
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Sign out
      description: |
        Ends the session of the Bearer token sent, if any; the client should
        discard the token. Always succeeds.
      operationId: logout
      security:
        - {}
        - bearerAuth: []
      responses:
        "204":
          description: Signed out.
  /api/auth/me:
    get:
      tags: [auth]
      summary: Get the signed-in user
      operationId: getCurrentUser
      responses:
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/users:
    get:
      tags: [users]
      summary: List users
      description: Sorted by username.
      operationId: listUsers
      responses:
        "200":
          description: All users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [users]
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: The created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The username is already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: username is already taken
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    put:
      tags: [users]
      summary: Update a user
      description: |
        Sets the administrator role and optionally a new password. Changing the
        password ends all of the user's sessions. Administrators cannot remove
        their own administrator role.
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [users]
      summary: Delete a user
      description: Administrators cannot delete their own account.
      operationId: deleteUser
      responses:
        "204":
          description: The user was deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}/tokens:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      summary: List a user's API tokens
      description: >-
        The user's unexpired API tokens, newest first. Sign-in sessions are
        not included, and token values are never returned again.
      operationId: listUserTokens
      responses:
        "200":
          description: The user's API tokens.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [users]
      summary: Create an API token for a user
      description: |
        Issues a named access token that authenticates as the user until it
        expires, e.g. for a device or integration. The lifetime is chosen
        freely between one minute and ten years. The token ends early when it
        is revoked, the user is deleted or their password is changed. Its
        value is only returned in this response.
      operationId: createUserToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: The token was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedApiToken"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}/tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - name: tokenId
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    delete:
      tags: [users]
      summary: Revoke an API token
      description: The token stops working immediately.
      operationId: revokeUserToken
      responses:
        "204":
          description: The token was revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such API token for this user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: token not found
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}/api-keys:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      summary: List a user's API keys
      description: Newest first.
      operationId: listUserApiKeys
      responses:
        "200":
          description: The user's API keys.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [users]
      summary: Register an API key for a user
      description: |
        Registers a public key under a caller-chosen ID. Whoever holds the
        private key can then act as the user by signing short-lived RS256
        JWTs with `kid` set to the ID (see the API description). To let
        another instance sync units from this one, use its server UUID and
        public key from its `GET /api/sync/identity`.
      operationId: createUserApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: The key was registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "409":
          description: An API key with this ID already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}/api-keys/{keyId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - name: keyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [users]
      summary: Delete an API key
      description: JWTs signed with it stop working immediately.
      operationId: deleteUserApiKey
      responses:
        "204":
          description: The key was deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such API key for this user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/groups:
    get:
      tags: [groups]
      summary: List groups
      description: >-
        Sorted by name, members by username. Groups are created when a member
        first signs in via SSO and kept after their last member leaves.
      operationId: listGroups
      responses:
        "200":
          description: All groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/settings:
    get:
      tags: [settings]
      summary: Get settings
      operationId: getSettings
      responses:
        "200":
          description: The current settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [settings]
      summary: Update settings
      description: >-
        Administrators only. Replaces all settings; an empty value restores the
        default. A change is recorded in the audit log as `settings.update`.
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: The settings as stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sync/identity:
    get:
      tags: [sync]
      summary: Get this instance's sync identity
      description: >-
        The server UUID and public key another instance registers as an API
        key to let this one sync from it. Both are generated on first start
        and stored in `KEYS_DIR`.
      operationId: getSyncIdentity
      responses:
        "200":
          description: This instance's identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/sync/remotes:
    get:
      tags: [sync]
      summary: List sync remotes
      description: Sorted by name.
      operationId: listSyncRemotes
      responses:
        "200":
          description: All sync remotes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyncRemote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [sync]
      summary: Add a sync remote
      description: An enabled remote starts syncing right away.
      operationId: createSyncRemote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRemoteRequest"
      responses:
        "201":
          description: The remote was added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRemote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sync/remotes/{id}:
    parameters:
      - $ref: "#/components/parameters/SyncRemoteId"
    get:
      tags: [sync]
      summary: Get a sync remote
      operationId: getSyncRemote
      responses:
        "200":
          description: The remote.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRemote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/SyncRemoteNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [sync]
      summary: Update a sync remote
      description: >-
        Its worker restarts when the URL or interval changes, and stops or
        starts with `enabled`.
      operationId: updateSyncRemote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRemoteRequest"
      responses:
        "200":
          description: The updated remote.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRemote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/SyncRemoteNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [sync]
      summary: Delete a sync remote
      description: Stops syncing; the units mirrored from it stay as ordinary local units.
      operationId: deleteSyncRemote
      responses:
        "204":
          description: The remote was deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/SyncRemoteNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/sync/remotes/{id}/trigger:
    parameters:
      - $ref: "#/components/parameters/SyncRemoteId"
    post:
      tags: [sync]
      summary: Start a full sync now
      operationId: triggerSyncRemote
      responses:
        "202":
          description: A full sync was queued.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/SyncRemoteNotFound"
        "409":
          description: The remote is disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/sync/remotes/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/SyncRemoteId"
    get:
      tags: [sync]
      summary: Get a sync remote's recent activity
      description: Oldest first, at most 200 entries; kept in memory only, so empty after a restart.
      operationId: listSyncRemoteLogs
      responses:
        "200":
          description: Recent log entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyncLogEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/SyncRemoteNotFound"
  /api/audit-log:
    get:
      tags: [audit]
      summary: List audit log entries
      description: >-
        Newest first. Records sign-ins (also failed ones), sign-outs and every
        change to users, API tokens and units. A unit's position is never
        recorded, including removing it, since the position history covers
        it; an update that changes only the position leaves no entry. Page through
        older entries by passing the last entry's `id` as `before`.
      operationId: listAuditLog
      parameters:
        - name: limit
          in: query
          description: Maximum number of entries to return.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: before
          in: query
          description: Only entries older than the entry with this ID.
          schema:
            type: integer
            minimum: 1
        - name: action
          in: query
          description: Only entries of this action.
          schema:
            $ref: "#/components/schemas/AuditAction"
        - name: actor
          in: query
          description: Only entries of the user with this ID.
          schema:
            type: integer
            minimum: 1
        - name: targetType
          in: query
          schema:
            $ref: "#/components/schemas/AuditTargetType"
        - name: targetId
          in: query
          description: Only entries about the record with this ID, e.g. a unit's UUID.
          schema:
            type: string
        - name: since
          in: query
          description: Only entries recorded at or after this time (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Only entries recorded at or before this time (RFC 3339). Must not be before `since`.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: The matching entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/units:
    get:
      tags: [units]
      summary: List units
      description: Sorted by name.
      operationId: listUnits
      responses:
        "200":
          description: All units.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [units]
      summary: Create a unit
      operationId: createUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitRequest"
      responses:
        "201":
          description: The created unit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/UnitNameTaken"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/units/events:
    get:
      tags: [units]
      summary: Stream unit changes
      description: |
        WebSocket endpoint (`ws://` / `wss://`). After the upgrade the server
        sends one JSON text message, a `UnitEvent`, for every unit that is
        created, updated or deleted; it ignores anything the client sends.

        Browsers can't set headers on a WebSocket, so besides the
        Authorization header the token is accepted as the subprotocols
        `bearer, <token>`; the server then selects `bearer`. Upgrades from
        any origin are accepted. The stream closes when the
        session ends (code 1008), when the server shuts down (1001), or when
        the client falls too far behind (1013). Changes made while
        disconnected are not replayed, so clients should refetch
        `GET /api/units` after reconnecting.
      operationId: streamUnitEvents
      responses:
        "101":
          description: Switched to the WebSocket protocol; messages are `UnitEvent`s.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An access token from `POST /api/auth/login`, the SSO flow or
        `POST /api/users/{id}/tokens`, or, when enabled, a JWT access token
        from a trusted OIDC provider.
...
  responses:
    BadRequest:
      description: The request body is malformed or fails validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid request body
    Unauthorized:
      description: No valid access token.
      headers:
        WWW-Authenticate:
          description: '`Bearer`, or `Bearer error="invalid_token"` when a token was sent but is invalid, expired or revoked.'
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
    Forbidden:
      description: The signed-in user is not an administrator.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: forbidden
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: internal error
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required: [token, tokenType, expiresAt, user]
      properties:
        token:
          type: string
          description: "JWT access token, sent as `Authorization: Bearer <token>`."
        tokenType:
          type: string
          const: Bearer
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...